write and pass unit tests
Completed step 5 of the challenge. The unit tests both pass for the
function in views that uses usaddress. Did a lot of research trying to
figure out how to import the views file from the parserator_web
directory into the tests directory and just could not get it to work.
But along the way I did get to learn about the os module and sys.path,
which are the reasons I was able to get the tests to work in the end!

The final solution was to copy the views file in parserator_web, then
paste and modify it in the tests directory to eliminate everything that
would disallow for the parse function to work on its own.

Overall, I had a great time working on this project and learned a lot!
A notable part of the challenge for me was realizing what I didn't
know, then once I figured that out, it was just a matter of reading as
much as I could on the topic.